+++ app/domains/agent/controller/agent_router.py
@@ -1,6 +1,6 @@
 from fastapi import APIRouter, Depends, HTTPException, status
 
-from app.domains.agent.repository.repository import ChatSessionRepository
+from app.domains.agent.repository.chat_session_repository import ChatSessionRepository
 from app.domains.agent.schema.schemas import (
     AgentChatRequest,
     AgentChatResponse,
