+++ app/domains/content/dependency/dependencies.py
@@ -1,4 +1,4 @@
-from app.domains.content.repository.repository import MongoContentCatalogRepository
+from app.domains.content.repository.catalog_repository import MongoContentCatalogRepository
 from app.domains.content.service.catalog_service import ContentCatalogService
 from app.infrastructure.db.mongo.mongo_client import get_mongo_client
