+++ tests/agent/test_session_service.py
@@ -5,7 +5,7 @@ from unittest.mock import AsyncMock, MagicMock
 import pytest
 
 from app.domains.agent.models import AgentDecision, ChatMessage, ChatSession
-from app.domains.agent.repository.repository import ChatSessionRepository
+from app.domains.agent.repository.chat_session_repository import ChatSessionRepository
 from app.domains.agent.service.agent_service import AgentService, ChatSessionService
 from app.domains.progress.models import StudentWeaknessProfile, WeakConcept
