+++ tests/structure/test_domain_structure.py
@@ -10,7 +10,7 @@ def test_domain_packages_are_importable():
         "app.domains.developer.admin_router",
         "app.domains.auth",
         "app.domains.auth.models",
-        "app.domains.auth.repository.repository",
+        "app.domains.auth.repository.auth_repository",
         "app.domains.auth.controller.auth_router",
         "app.domains.auth.schema.schemas",
         "app.domains.auth.service.auth_service",
@@ -19,7 +19,7 @@ def test_domain_packages_are_importable():
         "app.domains.progress.controller.teacher_student_router",
         "app.domains.classroom.dependency.dependencies",
         "app.domains.classroom.models",
-        "app.domains.classroom.repository.repository",
+        "app.domains.classroom.repository.classroom_repository",
         "app.domains.classroom.service.classroom_service",
         "app.domains.classroom.schema.schemas",
         "app.domains.agent.schema.schemas",
@@ -29,7 +29,7 @@ def test_domain_packages_are_importable():
         "app.domains.instruction.service.generation",
         "app.domains.instruction.validation.validation",
         "app.domains.instruction.models",
-        "app.domains.instruction.repository.repository",
+        "app.domains.instruction.repository.assignment_repository",
         "app.domains.instruction.schema.schemas",
         "app.domains.instruction.service.instruction_service",
         "app.domains.content.controller.catalog_router",
@@ -43,11 +43,11 @@ def test_domain_packages_are_importable():
         "app.domains.content.stage3.router",
         "app.domains.content",
         "app.domains.content.models",
-        "app.domains.content.repository.repository",
+        "app.domains.content.repository.catalog_repository",
         "app.domains.content.schema.schemas",
         "app.domains.content.service.catalog_service",
         "app.domains.progress.models",
-        "app.domains.progress.repository.repository",
+        "app.domains.progress.repository.learning_record_repository",
         "app.domains.progress.schema.schemas",
         "app.domains.progress.service.learning_record_service",
         "app.domains.content.stage3.schemas",
